# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "gestioneventos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/asistentes": {
            "get": {
                "tags": [
                    "Asistentes"
                ],
                "summary": "Mostrar todos los asistentes",
                "responses": {
                    "200": {
                        "description": "Lista de asistentes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Asistente"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron asistentes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Asistentes"
                ],
                "summary": "Crear un nuevo asistente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Asistente"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Asistente creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Asistente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de los datos"
                    },
                    "404": {
                        "description": "El evento o la ubicación no existen"
                    }
                }
            }
        },
        "/api/asistentes/{id}": {
            "get": {
                "tags": [
                    "Asistentes"
                ],
                "summary": "Mostrar un asistente específico",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asistente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Asistente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asistente no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Asistentes"
                ],
                "summary": "Actualizar un asistente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Asistente"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Asistente actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Asistente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de los datos"
                    },
                    "404": {
                        "description": "Asistente no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Asistentes"
                ],
                "summary": "Eliminar un asistente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asistente eliminado"
                    },
                    "404": {
                        "description": "Asistente no encontrado"
                    }
                }
            }
        },
        "/api/asistentes/{id}/partial": {
            "patch": {
                "tags": [
                    "Asistentes"
                ],
                "summary": "Actualizar parcialmente un asistente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_asistente": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "ci_asistente": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Asistente actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Asistente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de los datos"
                    },
                    "404": {
                        "description": "Asistente no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Asistente": {
                "required": [
                    "nombre_asistente",
                    "ci_asistente",
                    "telefono",
                    "id_evento",
                    "id_ubicacion"
                ],
                "properties": {
                    "id": {
                        "description": "ID del asistente",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre_asistente": {
                        "description": "Nombre del asistente",
                        "type": "string"
                    },
                    "ci_asistente": {
                        "description": "Cédula de identidad del asistente",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "Teléfono del asistente",
                        "type": "string"
                    },
                    "id_evento": {
                        "description": "ID del evento relacionado",
                        "type": "integer",
                        "format": "int64"
                    },
                    "id_ubicacion": {
                        "description": "ID de la ubicación relacionada",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}